expose llm types
