+++ packages/core/src/agent/index.ts
@@ -2,7 +2,8 @@ import { CoreMessage, UserContent } from 'ai';
 import { Integration } from '../integration';
 import { createLogger, Logger } from '../logger';
 import { AllTools, ToolApi } from '../tools/types';
-import { LLM, ModelConfig } from '../llm';
+import { LLM } from '../llm';
+import { ModelConfig } from '../llm/types';
 
 export class Agent<
   TIntegrations extends Integration[] | undefined = undefined,
